requirements.txt
haystack-ai
assemblyai